+++ lib/generators/active_agent/dashboard/install/templates/migrations/create_active_agent_sandbox_runs.rb
@@ -17,7 +17,7 @@ class CreateActiveAgentSandboxRuns < ActiveRecord::Migration<%= migration_versio
       t.datetime :completed_at
 
       # Screenshots
-      t.jsonb :screenshots, default: []
+      t.json :screenshots, default: []
 
       t.timestamps
     end